+++ packages/webawesome/src/components/dropdown-item/dropdown-item.ts
@@ -2,7 +2,9 @@ import type { PropertyValues } from 'lit';
 import { html } from 'lit';
 import { customElement, property, query, state } from 'lit/decorators.js';
 import { animateWithClass } from '../../internal/animate.js';
+import { warnDeprecatedSize } from '../../internal/size.js';
 import { HasSlotController } from '../../internal/slot.js';
+import { watch } from '../../internal/watch.js';
 import WebAwesomeElement from '../../internal/webawesome-element.js';
 import '../icon/icon.js';
 import styles from './dropdown-item.styles.js';
@@ -48,7 +50,12 @@ export default class WaDropdownItem extends WebAwesomeElement {
   /**
    * @internal The dropdown item's size.
    */
-  @property({ reflect: true }) size: 'small' | 'medium' | 'large' = 'medium';
+  @property({ reflect: true }) size: 'xs' | 's' | 'm' | 'l' | 'xl' | 'small' | 'medium' | 'large' = 'm';
+
+  @watch('size')
+  handleSizeChange() {
+    warnDeprecatedSize(this.localName, this.size);
+  }
 
   /**
    * @internal The controller will set this property to true when at least one checkbox exists in the dropdown. This